chore: add warnings about app contracts being unaudited

+++ account-age/src/AccountAge.sol
@@ -1,4 +1,7 @@
 // SPDX-License-Identifier: MIT
+// WARNING! This smart contract has not been audited.
+// DO NOT USE THIS CONTRACT FOR PRODUCTION
+// This is an example contract to demonstrate how to integrate an application with the audited production release of AxiomV1 and AxiomV1Query.
 pragma solidity 0.8.19;
 
 import {IAxiomV1Query} from "axiom-contracts/contracts/interfaces/IAxiomV1Query.sol";
@@ -8,6 +11,7 @@ contract AccountAge is Ownable {
     address public axiomQueryAddress;
 
     mapping(address => uint32) public birthBlocks;
+
     event UpdateAxiomQueryAddress(address newAddress);
     event AccountAgeVerified(address account, uint32 birthBlock);
 
@@ -16,27 +20,21 @@ contract AccountAge is Ownable {
         emit UpdateAxiomQueryAddress(_axiomQueryAddress);
     }
 
-    function updateAxiomQueryAddress(
-        address _axiomQueryAddress
-    ) external onlyOwner {
+    function updateAxiomQueryAddress(address _axiomQueryAddress) external onlyOwner {
         axiomQueryAddress = _axiomQueryAddress;
         emit UpdateAxiomQueryAddress(_axiomQueryAddress);
     }
 
-    function verifyAge(
-        IAxiomV1Query.AccountResponse[] calldata accountProofs,
-        bytes32[3] calldata keccakResponses
-    ) external {
+    function verifyAge(IAxiomV1Query.AccountResponse[] calldata accountProofs, bytes32[3] calldata keccakResponses)
+        external
+    {
         require(accountProofs.length == 2, "Too many account proofs");
         address account = accountProofs[0].addr;
         require(account == accountProofs[1].addr, "Accounts are not the same");
-        require(
-            accountProofs[0].blockNumber + 1 == accountProofs[1].blockNumber,
-            "Block numbers are not consecutive"
-        );
+        require(accountProofs[0].blockNumber + 1 == accountProofs[1].blockNumber, "Block numbers are not consecutive");
         require(accountProofs[0].nonce == 0, "Prev block nonce is not 0");
         require(accountProofs[1].nonce > 0, "No account transactions in curr block");
-        uint addrSize;
+        uint256 addrSize;
         assembly {
             addrSize := extcodesize(account)
         }
